feat(api/audio): add tagging to spectrogram, track api changes
This includes a several changes to the UI to allow for tagging functionality, and the ability to add and delete tracks. There are additional api ends including ability to get a single track which is more useful on the audio side where you make only have a few tracks. As well as including making positions for tracks normalized and include generic language (frame_num -> order).