feat(audio): add immer, change colours, add mute icon, major refactor audio, typescript changes
This commit includes a number of larger changes including the use of composition api and immer along with styling and functionality changes on audio side. The reason for using composition is to support much better typing, with Volar also allowing for future Vue 3 use. Immer to implement immutable objects which allow for state management across time which is particularly useful to allow for redo/undo, which audio has more of (add/deleting/modifying tracks, along with the normal tags)