Audio list info (#40)
* refactor(Chart.js/Analysis): upgrade chart.js to 3.6.2, fix undefined usage, == to ===

Clean up in prep for using more performant 3.6, with migration/clean up along the way

* feat(scripts): add deno scripts to add user, device, and recordings

Primary reason for using deno is to remove the need for more modules and speed to compile. These are light scripts to call and can be found here https://github.com/denoland/deno. They integrate with the existing types and don't require additional tsc scripts to run.

* feat(Audio): migrate additionalMetadata to own columns & tracks, add audio tags & cacophony index graph

Using sequelize to migrate data for audio recordings in Recordings table, taking cacophony_index to own cacophonyIndex column, audio Analysis(morepork tags) to Track & tags, and added corresponding UI elements to the Recording List

* fix(tests/audio): change cacophony index & analysis properties in tests, change types

This changes aims to access the new location of the cacophony index and analysis for audio with some fixes around reporting & accessing the new cacophonyIndex

* style(browse): automatic lint

* filter false positive tracks from ui

* comment

* margine

* re add template

* refactor(migration): change use of sequelize models to use only SQL queries

Was able to access the Models in the Travis CI, so reverted to only SQL. Generally not a good idea to use the models in migrations anyhow.

* fix(migration): check for algorithm in DetailSnapshots

* try render frame

* lint fix

* fix(migration): use replacements in sequelize migration, add automatic to Track table

Adding automatic to Track table as audio will have the ability to have manual tracks being added. It would be useful therefore to be able to distinguish them in a single query.

* add filtering to recordings view

* ran prettier

* feat(Audio): add Track List UI, Spectrogram, Track Selection, fix migration

This is a rough implementation of parts of the Audio UI along side a fix for migration

* fix human disagree logic

* fix(migration): check for cacophonyIndex

* fix(migration): default value for liklihood

Ran into issue where somes Tags don't have a liklihood

* fix(migration): reverse boolean check, add default value

* fix master tags

* linted

* updated

* updated styles and fixed sticky

* add filtered column

* fix required

* update filtered

* feat(api/audio): add tagging to spectrogram, track api changes

This includes a several changes to the UI to allow for tagging functionality, and the ability to add and delete tracks. There are additional api ends including ability to get a single track which is more useful on the audio side where you make only have a few tracks. As well as including making positions for tracks normalized and include generic language (frame_num -> order).

* fixed up some tests

* fix(lint): recording lint

* fix(recordings): register cacophony index graph

* fix(merge): vue reactivity, api endpoints fix, add tracks/tags recording list

Several issues occured with merge, included fixes as well as reimplemententing the track & tags

* fix(audio): add track data type to track request, basic confirm

* - Remove DeviceUsers from api and ui.
- Delete a whole lot of dead code from the api.

* fix(audio): allow backwards tracks, fix delete, add float to track pos

* refactor(scripts): remove personal scripts

* progressing deviceuser removal from tests

* all device users removed from tests

* deleted test suites for endpoints we no longer support

* - Refactoring recordings query

* - Fix monitoring query

* feat(Vue): add composition-api, vue 2 compatability

This is a future proofing changing in case of moving to Vue 3. This way any new components can be made in a way that lessens migration issues. It also allows the use of Volar vue tooling.

* dont skip the test

* - Fixing monitoring tests

* - Fixing monitoring tests

* - Fix duplicate visits timing issue.

* - Fix group location map link on groupsview.

* clean up merge

* finished merging

* add back in additional meta

* add filtered

* update to work with separate tracks

* update filtered logic when no ai tags

* skip travis for now

* update api doc

* feat(audio): add immer, change colours, add mute icon, major refactor audio, typescript changes

This commit includes a number of larger changes including the use of composition api and immer along with styling and functionality changes on audio side. The reason for using composition is to support much better typing, with Volar also allowing for future Vue 3 use. Immer to implement immutable objects which allow for state management across time which is particularly useful to allow for redo/undo, which audio has more of (add/deleting/modifying tracks, along with the normal tags)

* fix no tags filtering and reprocess filtering

* set filtered on reprocess

* use method

* update filtered with transaction

* refactor(audio): functions to utils, add immutable state, style changes

This adds a useState to modify the reactive state of a ref with a proper immutable structure

* Add filter tests, rework to use data from commands/dataTemplates

* Add filter tests, rework to use data from commands/dataTemplates - bugfixes

* Add filter tests, rework to use data from commands/dataTemplates - linting

* run tests, fix empty track not being filtered

* update filtered from fileprocessing api

* fixed lock file

* capital M for master on tests

* fix some tests

* remove only

* filtered true

* completed filtering tests

* centralise filtered_tags definition

* tests - templates - stop tests from modifying global test-data templates

* fix z index

* feat(audio): undo functionality on tracks

Reuses the existing track in memory to add the track and tags

* feat(audio): add bird labels, playlist w/ nav, delete, download, useRoute/r, styling

* fix(audio): delete unused, add frequency to track, fix colours, and common bird buttons

* fix(audio): add freq from merge, disregard temp track colour

* fix(test): change frame_number to order

* fix(test): change cacophony index for test

* fix(audio): add cacophony index (tooltips), alphabetical list, check touch events available

* fix(audio): remove check for interface

* fix(recordings): check type is available in row view

* fix(recording): change props key, remove type column

* merge in combined intergration tests

* Finalised tests to run against test server (linting)

* fix(recordings): change z-index on sidebar on smaller screen

* fix(migration): remove Promise.all due to memory heap

Co-authored-by: gferraro <g.ferraro22@gmail.com>
Co-authored-by: jon <jon@hardiesoft.com>
Co-authored-by: gw7nvw <mattbriggs@yahoo.com>